\name{ctglm.run}
\alias{ctglm.run}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
ctglm model run
}
\description{
run ctglm model
}
\usage{
ctglm.run ( s, work.dir=getwd(), verbose=TRUE, ... )
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{s}{ctglm syntax object (from ctglm.syntax)}
  \item{work.dir}{working directory}
  \item{verbose}{output messages to console}
  \item{...}{additional parameters}
}
%% \details{}
\value{
returns a ctglm results list
}
%% \references{}
\author{
Martin Hecht
}
%% \note{}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\section{Warning}{
This version is alpha. Use with care.
} 

\seealso{
}

\examples{
}
