\name{ctglm.save.syntax}
\alias{ctglm.save.syntax}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
save ctglm syntax
}
\description{
save ctglm syntax
}
\usage{
ctglm.save.syntax ( s, dir, ... )
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{s}{syntax/call object (from ctglm.syntax)}
  \item{dir}{directory for output}
  \item{...}{additional parameters}
}
%% \details{}
\value{
saves syntax/call/data, returns TRUE
}
%% \references{}
\author{
Martin Hecht
}
%% \note{}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\section{Warning}{
This version is alpha. Use with care.
} 

\seealso{
}

\examples{
}
