\name{ctglm.syntax}
\alias{ctglm.syntax}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
ctglm syntax/call creation
}
\description{
create ctglm syntax/call
}
\usage{
ctglm.syntax ( m, model.name="model", ..., verbose=TRUE )
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{m}{ctglm model object (from ctglm.model)}
  \item{model.name}{name of the model}
  \item{...}{additional arguments. jags.add: additional jags syntax (must be 1-column matrix)}
  \item{verbose}{output messages to console}
}
%% \details{}
\value{
returns a ctglm syntax list
}
%% \references{}
\author{
Martin Hecht
}
%% \note{}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\section{Warning}{
This version is alpha. Use with care.
} 

\seealso{
}

\examples{
}
