\name{mcarray.chains.combine}
\alias{mcarray.chains.combine}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
combine chains of mcarray objects
}
\description{
combine chains of mcarray objects
}
\usage{
mcarray.chains.combine ( x )
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{list (chains) of mcarray lists}
}
%% \details{}
\value{
returns mcarray list
}
%% \references{}
\author{
Martin Hecht
}
%% \note{}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\section{Warning}{
This version is alpha. Use with care.
} 

\seealso{
}

\examples{
}
